backend/img_service/pyproject.toml
[project]
name = "img-service"
version = "0.1.0"
description = "FastAPI image utility service for Catdex assets"
authors = [{name = "BeastyRabbit Tools"}]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.2",
  "uvicorn[standard]>=0.30.6",
  "httpx>=0.27.2",
  "pillow>=10.4.0",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.4.2", "httpx>=0.27.2", "pytest-asyncio>=0.24.0"]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["img_service*"]
